# docs/openapi.yml
...
openapi: 3.1.0
components:
  schemas:
...
    ApiError:
      type: object
      description: "The error envelope: a message, plus a reason or a code on a 409."
      properties:
        message:
          type: string
          description: "What went wrong, in words. Always present."
        reason:
          $ref: "#/components/schemas/Reason"
          type: string
          description: "Which refusal this is. Present on the 409s the release and\
            \ integrate flows raise, absent on every other error. A value a client\
            \ does not recognise is shown as the message verbatim."
        conflicts:
          type: array
          items:
            type: string
          description: The conflicted paths. Present only for CONFLICT and MERGE_CONFLICT.
        code:
          type: string
          description: "The refusal's own name, where it has one (e.g. RUNNER_OWNS_WORKSPACES).\
            \ Absent otherwise."
        workspaceIds:
          type: array
          items:
            type: integer
            format: int64
          description: The ACTIVE workspace row ids a runner still owns. Present only
            with RUNNER_OWNS_WORKSPACES.
        workspaces:
          type: array
          items:
            $ref: "#/components/schemas/RunnerOwnedWorkspace"
          description: "The ACTIVE workspaces a runner still owns, each with its repository\
            \ and branch so a client can link it. Present only with RUNNER_OWNS_WORKSPACES."
...
    CleanupBranchRequest:
      type: object
      required:
      - branch
      properties:
        branch:
          type: string
          pattern: \S
        result:
          type: string
...
    MergeBranchRequest:
      type: object
      required:
      - source
      properties:
        source:
          type: string
          pattern: \S
        target:
          type: string
        result:
          type: string
...
    Reason:
      type: string
      enum:
      - CONFLICT
      - MERGE_CONFLICT
      - NOT_FAST_FORWARD
      - ALREADY_INTEGRATED
      - PUSH_REJECTED
      - RELEASE_REQUIRED
...
    Response1:
      type: object
      properties:
        commitHash:
          type: string
        hasConflicts:
          type: boolean
        output:
          type: string
        cleanedUp:
          type: boolean
...
    Response2:
      type: object
      properties:
        success:
          type: boolean
...
    RunnerOwnedWorkspace:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: The workspace row id
        repositoryId:
          type: string
          description: Its repository
        branch:
          type: string
          description: "Its branch, or null"
...
  securitySchemes:
    SecurityScheme:
      type: openIdConnect
      openIdConnectUrl: http://dev-qits-idp:8080/idp/.well-known/openid-configuration
      description: Authentication
paths:
...
  /workspaces/api/branches/cleanup:
    post:
      parameters:
      - name: repositoryId
        in: query
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CleanupBranchRequest"
        required: true
      responses:
        "200":
          description: Removed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response2"
        "400":
          description: "The branch is not eligible: uncommitted changes, unmerged\
            \ commits, or dependents."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: No such repository.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Cleanup Branch
      tags:
      - Branch Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
  /workspaces/api/branches/merge:
    post:
      parameters:
      - name: repositoryId
        in: query
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MergeBranchRequest"
        required: true
      responses:
        "200":
          description: Merged.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response1"
        "400":
          description: "A blank, dash-leading or self-referential branch name."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: No such repository.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: The target is the repository's default branch. `reason` is
            RELEASE_REQUIRED and the message names the release flow that does write
            it.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Merge Branch
      tags:
      - Branch Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
...
  /workspaces/api/editor/ensure:
    post:
      operationId: ensureEditor
      responses:
        "201":
          description: The editor was started by this call.
        "200":
          description: The editor was already there.
          content:
            application/json:
              schema: {}
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      summary: Ensure
      tags:
      - Editor Controller
      security:
      - SecurityScheme:
        - qits:admin
        - qits:admin-agent
...
info:
  title: qits-workspaces API
  version: 2026.1005.203201